Cleanup message interfaces
