+++ packages/client-p2p-messages/src/types.d.ts
@@ -6,10 +6,11 @@ import BN from 'bn.js';
 import { AccountId, BlockNumber, HeaderHash, ParaChainId, Signature } from '@polkadot/primitives/base';
 import { Justification } from '@polkadot/primitives/bft';
 import { Header } from '@polkadot/primitives/header';
+import { JsonHeader } from '@polkadot/primitives/json/types';
 import { Role } from '@polkadot/primitives/role';
 
 export interface MessageEncoder <M> {
-  type: number;
+  readonly type: number;
   encode (): M;
 }
 
@@ -20,9 +21,26 @@ export interface MessageDecoder <M, C> {
   decode (input: M): C;
 }
 
+export type BlockAnnounceEncoded = {
+  BlockAnnounce: {
+    header: JsonHeader
+  }
+};
+
 export interface BlockAnnounceMessage {
   header: Header
 }
+
+export type BlockRequestEncoded = {
+  BlockRequest: {
+    direction: BlockRequestMessageDirection,
+    fields: Array<BlockRequestMessageField>,
+    from: string | number,
+    id: number,
+    max: number,
+    to?: string | number | null
+  }
+};
 
 export type BlockRequestMessageField = 'Header' | 'Body' | 'Receipt' | 'MessageQueue' | 'Justification';
 
@@ -37,6 +55,22 @@ export type BlockRequestMessage = {
   to?: HeaderHash | null
 }
 
+export type BlockResponseEncodedBlock = {
+  hash: string,
+  header: Header,
+  body: Array<Array<number>>,
+  receipt: null,
+  messageQueue: null,
+  justification: Justification
+};
+
+export type BlockResponseEncoded = {
+  BlockResponse: {
+    id: number,
+    blocks: Array<BlockResponseEncodedBlock>
+  }
+};
+
 export type BlockResponseMessageBlock = {
   hash: Uint8Array,
   header: Header,
@@ -49,6 +83,19 @@ export type BlockResponseMessage = {
   id: number,
   blocks: Array<BlockResponseMessageBlock>,
 }
+
+export type StatusEncoded = {
+  Status: {
+    best_hash: string,
+    best_number: number,
+    genesis_hash: string,
+    parachain_id?: string | null,
+    roles: Array<string>,
+    validator_id?: string | null,
+    validator_signature?: string | null,
+    version: number
+  }
+};
 
 export interface StatusMessage {
   bestHash: HeaderHash,
